Added the Header, Footer, Hero, and Layout component.
I created the desktop version for the Header component that includes the logo, the nav links, and the contact button. Also, created the Hero section (component) that will the first thing that new visitants will see and It will show which is my current job. Then I started creating the developing process section. For this, I created the ProcessStep component that will be used to show the four steps that I use during the creation of new projects.